feat: expose capnpc-ts, capnpc-dts, capnpc-js
