+++ build.config.ts
@@ -5,7 +5,9 @@ export default defineBuildConfig({
   entries: [
     "./src/index.ts",
     "./src/compiler/index.ts",
-    "./src/compiler/main.ts",
+    "./src/compiler/capnpc-js.ts",
+    "./src/compiler/capnpc-dts.ts",
+    "./src/compiler/capnpc-ts.ts",
     ...["cpp", "persistent", "rpc-twoparty", "rpc", "schema", "ts"].map(
       (n) => `./src/capnp/${n}.ts`,
     ),
