+++ src/compiler/capnpc-dts.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { main } from "./capnp-es";
+
+await main("dts");
