+++ src/compiler/capnpc-ts.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { main } from "./capnp-es";
+
+await main("ts");
